requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers
python-dotenv
pypdf
plotly
pydeck
pandas
scikit-learn